Factor 'by-recap-customer-codes' out from the nypl-core-objects test
